requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
black==24.4.2
    # via -r requirements.in
build==1.2.1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
mypy==1.11.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.1
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via black
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via mypy
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
